Split waveformat to own file
